requirements.txt
pandas==2.2.3
numpy==2.2.4
requests==2.32.3
sqlalchemy==2.0.39
pyarrow==19.0.1
streamlit==1.44.1
plotly==6.0.1
python-dotenv==1.0.1
pydantic==2.10.6
pytest==8.3.5
